Made a litter model specific conftest.py file
